% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{assess.quality}
\alias{assess.quality}
\title{assess.quality}
\usage{
assess.quality(
  platformList,
  mlmList,
  fileList,
  numberOfIterations = 100,
  pretreatment = "no_pretreatment",
  percentageForTrainingSet = 0.7
)
}
\arguments{
\item{platformList}{platformList}

\item{mlmList}{mlmList}

\item{fileList}{fileList}

\item{numberOfIterations}{numberOfIterations}

\item{pretreatment}{pretreatment}

\item{percentageForTrainingSet}{percentageForTrainingSet}

\item{fileTabList}{fileTabList}
}
\value{

}
\description{
assess.quality
}
\examples{
\dontrun{assess.quality(platformList, mlmList, fileList, numberOfIterations, pretreatment, percentageForTrainingSet
)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
