% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticRegression.R
\name{elasticRegression.run}
\alias{elasticRegression.run}
\title{Main function to calculate performance of elastic regression model}
\usage{
elasticRegression.run(regressionParameterList)
}
\arguments{
\item{regressionParameterList}{a list which contains
number_of_iterations: number of iterations to calculate performance
pretreatment: data pretreatment method (auto-scale, mean-center or range-scale
is supported)
percentageForTrainingSet: percentage of samples in training dataset
dataSet: dataFrame which is read from data file and subjected to the model.}
}
\value{
a list containing performance results
RMSEList: a list which contains RMSE of each iteration
cumulativeRMSEList: a list which contains cumulative RMSE mean in
each iteration
RMSE: mean RMSE of all iterations
RSquareList: a list which contains RSquare of each iteration
cumulativeRSquareList : a list which contains cumulative RSquare mean in
each iteration
RSquare: mean RSquare of all iterations
bestHyperParamsList: a list containing best alpha and lambda values for each iteration
}
\description{
After pretreatment on dataset this function calculates performance
of elastic regression model through iterations and returns performance metrics.
In each iteration different partitioning is done on dataset to create
training and validation datasets, cross-validation tuning is done on training
dataset to find optimum lambda value
}
\examples{
\dontrun{elasticRegression.run(regressionParameterList)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
