% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalMetrics}
\alias{evalMetrics}
\title{evalMetrics}
\usage{
evalMetrics(true, predicted)
}
\value{
list containing RSquare and RMSE performance metric
}
\description{
generates RSQUARE and RMSE performance metric from predicted and
actual values
}
\examples{
\dontrun{evalMetrics(true, predicted)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
