% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{generatePCAPlots}
\alias{generatePCAPlots}
\title{generatePCAPlots}
\usage{
generatePCAPlots(dataSet, outputDir, platform)
}
\arguments{
\item{dataSet}{dataFrame from the data file generated by analytical platform}

\item{outputDir}{output directory name provided by the use in config json file}

\item{platform}{manalytical platform name}
}
\description{
generates PCA plot for each analytical platform
}
\examples{
\dontrun{generatePCAPlots(dataSet, outputDir, platform)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
