% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancePlots.R
\name{generatePerformancePlot}
\alias{generatePerformancePlot}
\title{generatePerformancePlot}
\usage{
generatePerformancePlot(data, plotName, dataNames, file)
}
\arguments{
\item{data}{cumulative performance metrics results}

\item{plotName}{RMSE Means OR rsquare Means}

\item{dataNames}{machine learning model names}

\item{file}{output pdf file name}
}
\description{
generates performance plot for each analytical platform
}
\examples{
\dontrun{generatePerformancePlot(data, plotName, dataNames, file)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
