% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancePlots.R
\name{generatePerformancePlots}
\alias{generatePerformancePlots}
\title{generatePerformancePlots}
\usage{
generatePerformancePlots(platformPerformanceResults, outputDir)
}
\arguments{
\item{outputDir}{}
}
\description{
generates performance plots for each platform. Plots show
RMSE and RSquare change in different number of iterations
}
\examples{
\dontrun{generatePerformancePlot(data, plotName, dataNames, file)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
