% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{generateStatistics}
\alias{generateStatistics}
\title{generateStatistics}
\usage{
generateStatistics(platformPerformanceResults, outputDir, createStatisticsFile)
}
\arguments{
\item{platformPerformanceResults}{list of machine learning performance results
for each platform}

\item{outputDir}{output directory name provided by the use in config json file}

\item{createStatisticsFile}{boolean value indicating whether to create statistics
file or not}
}
\description{
generates RMSE and RSQUARE statistics report for each platform
and machine learning models
}
\examples{
\dontrun{generateStatistics(platformPerformanceResults, outputDir, createStatisticsFile)}

}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
