% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gePretreatmentVector}
\alias{gePretreatmentVector}
\title{gePretreatmentVector}
\usage{
gePretreatmentVector(pretreatment)
}
\arguments{
\item{pretreatment}{pretreatment string object  provided by config file}
}
\value{
vector
}
\description{
converts string of pretreatment parameter to vector to be used
in data scaling before machine learning modeling
}
\examples{
\dontrun{gePretreatmentVector(pretreatment)}

}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
