% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRegressionParameters}
\alias{getRegressionParameters}
\title{getRegressionParameters}
\usage{
getRegressionParameters(mlm, dataSet, platform)
}
\arguments{
\item{mlm}{machine learning model list}

\item{dataSet}{dataFrame produced by analytical platforms}
}
\value{
list of regression parameters
}
\description{
creates  regressionParameterList which contains all relevant
information to be used by machine learning models. regressionParameterList
is used as an internal object passed through different functions in machine
learning modeling
}
\examples{
\dontrun{getRegressionParameters(mlm, dataSet, platform )}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
