% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetwork.R
\name{neuralNetwork.run}
\alias{neuralNetwork.run}
\title{Main function to calculate performance of neural network model}
\usage{
neuralNetwork.run(regressionParameterList)
}
\arguments{
\item{regressionParameterList}{a list which contains
number_of_iterations: number of iterations to calculate performance
pretreatment: data pretreatment method (auto-scale, mean-center or range-scale
is supported)
percentageForTrainingSet: percentage of samples in training dataset
dataSet: dataFrame which is read from data file and subjected to the model.}
}
\value{
a list containing performance metrics
RMSE: mean RMSE of all iterations
RSquare: mean RSquare of all iterations
}
\description{
After pretreatment on dataset this function calculates performance
of neural network model
}
\examples{
\dontrun{neuralNetwork.run(regressionParameterList)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
