% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.pca}
\alias{plot.pca}
\title{plot.pca}
\usage{
\method{plot}{pca}(dataAll)
}
\arguments{
\item{dataAll}{dataFrame object}
}
\value{

}
\description{
draws  pca and biplot for the given dataset
}
\examples{
\dontrun{plot.pca(dataAll)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
