% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPrediction}
\alias{plotPrediction}
\title{plotPrediction}
\usage{
plotPrediction(model, testSet)
}
\arguments{
\item{model}{machine learning model}

\item{testSet}{dataFrame used for  machine learning model validation}
}
\description{
draws plot of predicted values against actual values with RMSE value
}
\examples{
\dontrun{plotPrediction(model, testSet )}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
