% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionTree.R
\name{pruneTree}
\alias{pruneTree}
\title{Cost-complexity post pruning of regression treed}
\usage{
pruneTree(modelTree, testSet)
}
\arguments{
\item{modelTree}{regression tree model created on training set before pruning}

\item{testSet}{validation dataset}
}
\value{
a list which contains
RMSEList:  a list of RMSE values calculated for each cost-complexity
bestPrunedTree: the best pruned tree model found according to RMSE metric
RMSE: RMSE value of the best pruned tree
RSquare: RSquare value of the best pruned tree
bestCp: cost comlexity of the best pruned tree
}
\description{
This method post-prunes the regression tree by cost-complexiy
to overcome issue of over-fitting to training set
}
\examples{
\dontrun{pruneTree(modelTree, testSet )}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
