% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readConfigFile}
\alias{readConfigFile}
\title{readConfigFile}
\usage{
readConfigFile(configFile)
}
\arguments{
\item{configFile}{configFile}
}
\value{
list of config object containing machineLearningModels, outputDirectory,
createStatisticsFile, createPerformancePlots, createPCAPlots
}
\description{
reads, parses and performs some validity checks on the configuration data
supplied in json format as config file
}
\examples{
\dontrun{readConfigFile(configFile)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
