% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readDataset}
\alias{readDataset}
\title{readDataset}
\usage{
readDataset(dataFileName)
}
\arguments{
\item{dataFileName}{dataFileName}
}
\value{
dataFrame
}
\description{
reads, parses and performs some validity checks on the dataset.
Data manipulation is done if it is needed. For example if the number of features exceeds
200, it is reduced by feature selection. Any name of feature is numeric, it is converted to
character value as it is required in some machine learning models.
supplied in json format as config file
}
\examples{
\dontrun{readConfigFile(configFile)}

}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
