% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeRedundantFeatures}
\alias{removeRedundantFeatures}
\title{removeRedundantFeatures}
\usage{
removeRedundantFeatures(dataSet)
}
\arguments{
\item{dataSet}{dataFrame object}
}
\value{

}
\description{
reduces the number of features in the dataset by
removing the reduntant during the variable selection process of data modeling.
}
\examples{
new dataset with reduced number of features
\dontrun{removeRedundantFeatures(selectFeatures)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
