% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(true, predicted)
}
\value{
RMSE performance metric
}
\description{
generates RMSE performance metric from predicted and
actual values
}
\examples{
\dontrun{RMSE(true, predicted)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
