% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RSQUARE}
\alias{RSQUARE}
\title{RSQUARE}
\usage{
RSQUARE(true, predicted)
}
\value{
RSquare performance metric
}
\description{
generates RSQUARE performance metric from predicted and
actual values
}
\examples{
\dontrun{RSQUARE(true, predicted)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
