% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{run.analysis}
\alias{run.analysis}
\title{run.analysis}
\usage{
run.analysis(configParams)
}
\arguments{
\item{configParams}{list containing parameters provided in config json file
by the user}
}
\description{
calls each machine learning run methods according to
the method parameter in regressionParameterList
}
\examples{
\dontrun{run.analysis(configParams)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
