% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{run.regression}
\alias{run.regression}
\title{run.regression}
\usage{
run.regression(regressionParameterList)
}
\arguments{
\item{regressionParameterList}{list containing parameters needed by machine
learning models to run}
}
\value{
a list which contains performance results for machine learning
model
}
\description{
calls each machine learning run methods according to
the method parameter in regressionParameterList
}
\examples{
\dontrun{run.regression(regressionParameterList)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
