% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoost.R
\name{XGBoost.run}
\alias{XGBoost.run}
\title{Main function to calculate performance of XGBoost model}
\usage{
XGBoost.run(regressionParameterList)
}
\arguments{
\item{regressionParameterList}{a list which contains
number_of_iterations: number of iterations to calculate performance
pretreatment: data pretreatment method (auto-scale, mean-center or range-scale
is supported)
percentageForTrainingSet: percentage of samples in training dataset
dataSet: dataFrame which is read from data file and subjected to the model.}
}
\value{
a list containing performance metrics
RMSE: mean RMSE of all iterations
RSquare: mean RSquare of all iterations
}
\description{
This function calculates performance XGBoost(gradient boosting)
}
\examples{
\dontrun{XGBoost.run(regressionParameterList)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
