% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{assess.quality}
\alias{assess.quality}
\title{assess.quality}
\usage{
assess.quality(configFile = configFile)
}
\arguments{
\item{configFile}{configFile}
}
\value{

}
\description{
assess.quality
}
\examples{
\dontrun{assess.quality(configFile)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
