% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectFeatures}
\alias{selectFeatures}
\title{selectFeatures}
\usage{
selectFeatures(dataSet)
}
\arguments{
\item{dataSet}{dataFrame object}
}
\value{
dataFrame
}
\description{
reduces the number of features in the dataset
by selecting the more important fautures
}
\examples{
new dataset with reduced number of features
\dontrun{selectFeatures(dataSet)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
